requirements.txt
Flask==3.0.2
psycopg2==2.9.9
SQLAlchemy==2.0.27
PyDispatcher==2.0.7
